move toJSON from Level to LevelHelper to reduce the dependency of the persistence module on the Servlet API
TM-260, TM-139 verwijder Servlet API dependency uit persistence module

reset npm files, not part of this Pr
